fix: double quote variable with shasum output in echo
The output of shasum xx is "<sum>  xx" with 2 spaces inbetween the
shasum and the filename. If that output is passed to echo without double
quotes, echo will see this as 2 arguments and will print them with a
single space inbetween. By double quoting it, echo will simply output
the single argument as-is.